% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearOrdering.R
\name{clusterAnalysis}
\alias{clusterAnalysis}
\title{Create clusterAnalysis}
\usage{
clusterAnalysis(df)
}
\arguments{
\item{df}{input data}
}
\value{
clusterAnalysis
}
\description{
Create clusterAnalysis
}
