% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearOrdering.R
\name{plot.clusterAnalysis}
\alias{plot.clusterAnalysis}
\title{Plot clusterAnalysis}
\usage{
\method{plot}{clusterAnalysis}(x)
}
\arguments{
\item{x}{clusterAnalysis}
}
\description{
Plot clusterAnalysis
}
