% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearOrdering.R
\name{print.clusterAnalysis}
\alias{print.clusterAnalysis}
\title{Print clusterAnalysis}
\usage{
\method{print}{clusterAnalysis}(x)
}
\arguments{
\item{x}{clusterAnalysis}
}
\description{
Print clusterAnalysis
}
