% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearOrdering.R
\name{summary.clusterAnalysis}
\alias{summary.clusterAnalysis}
\title{Print summary of clusterAnalysis}
\usage{
\method{summary}{clusterAnalysis}(x)
}
\arguments{
\item{x}{clusterAnalysis}
}
\description{
Print summary of clusterAnalysis
}
